Remove AnyRef and AnyValue
