+++ Source/OCompiler/Builtins/Type/Boolean.cs
@@ -2,7 +2,7 @@ using OCompiler.StandardLibrary.Type.Reference;
 
 namespace OCompiler.StandardLibrary.Type.Value;
 
-public class Boolean : AnyValue
+public class Boolean : Class
 {
     public bool _value;
