+++ Source/OCompiler/Builtins/Type/Class.cs
