+++ Source/OCompiler/Builtins/Type/Void.cs
@@ -1,6 +1,6 @@
 ﻿namespace OCompiler.StandardLibrary.Type.Value;
 
-public class Void : AnyValue
+public class Void : Class
 {
     /*        ඞ        */
 }
