--- Source/OCompiler/StandardLibrary/Type/Reference/AnyRef.cs
@@ -1,5 +1,0 @@
-namespace OCompiler.StandardLibrary.Type.Reference;
-
-public abstract class AnyRef: Class
-{
-}
