--- Source/OCompiler/StandardLibrary/Type/Value/AnyValue.cs
@@ -1,6 +1,0 @@
-﻿namespace OCompiler.StandardLibrary.Type.Value;
-
-public abstract class AnyValue : Class
-{
-}
-
